% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyPeasyGenotype.R
\name{EasyPeasyGenotype}
\alias{EasyPeasyGenotype}
\title{EasyPeasyGenotype}
\usage{
EasyPeasyGenotype(file_name, differential_temperature, het_lowest_bump)
}
\arguments{
\item{file_name}{.txt file exported from lightcycler96 software, save column as a sample}

\item{differential_temperature}{set critical temp. between picks of WT, and -/-. in case of tsc2 it's 80.5}

\item{het_lowest_bump}{set threshold of Resolight value in temperature period of 76-78 C, where value is below heterozygous bump and above curves for WT and -/-}
}
\value{
print in console information about numbers of each genotype, function also give data_frame with plate wells and genotype assigned to the well. Return two plots of HRM curves, one plot for combined curves, and one for each curve separate
}
\description{
EasyPeasyGenotype
}
\examples{
EasyPeasyGenotype(file_name = tsc2_none.txt, differential_temperature = 80.5, het_lowest_bump = 0.115
}
